backend/requirements.txt
fastapi
uvicorn
pydicom
pillow
requests
inference-sdk
python-multipart
pylibjpeg
pylibjpeg-libjpeg
pylibjpeg-openjpeg
google-genai
python-dotenv
